Correct formatting of keywords.txt
The DATA_TYPE keyword identifier is only valid when used in the RSYNTAXTEXTAREA_TOKENTYPE field of keywords.txt. That is the fourth field rather than the second, where DATA_TYPE was previously. I moved DATA_TYPE to the fourth field and replaced it with the equivalent KEYWORD1 KEYWORD_TOKENTYPE in the second field.